Add separate parsers
This refactors the code to have separate parsers for Asset Bundles
and Addressables Build Layout files. The parser implements a
CanParse and Parse method to handle applicability. Currently
there's no overlap so we won't be double-parsing, but it's
possible in the future.
