--- Analyzer/SQLite/SQLiteWriter.cs
@@ -1,870 +1,0 @@
-using System;
-using System.Collections.Generic;
-using Microsoft.Data.Sqlite;
-using System.IO;
-using System.Text.RegularExpressions;
-using UnityDataTools.Analyzer.SQLite.Handlers;
-using UnityDataTools.FileSystem;
-using UnityDataTools.FileSystem.TypeTreeReaders;
-using UnityDataTools.Analyzer.Build;
-using Newtonsoft.Json;
-using Analyzer.SQLite.Commands.AddressablesBuildReport;
-using UnityDataTools.Analyzer.SQLite.Commands.SerializedFile;
-
-namespace UnityDataTools.Analyzer.SQLite;
-
-public class SQLiteWriter : IWriter
-{
-    private HashSet<int> m_TypeSet = new();
-
-    private int m_CurrentAssetBundleId = -1;
-    private int m_NextAssetBundleId = 0;
-
-    private string m_DatabaseName;
-    private bool m_SkipReferences;
-
-    private Util.IdProvider<string> m_SerializedFileIdProvider = new();
-    private Util.ObjectIdProvider m_ObjectIdProvider = new();
-
-    private Regex m_RegexSceneFile = new(@"BuildPlayer-([^\.]+)(?:\.sharedAssets)?");
-
-    // Used to map PPtr fileId to its corresponding serialized file id in the database.
-    Dictionary<int, int> m_LocalToDbFileId = new ();
-
-    // Used to map PPtr fileId to its corresponding serialized file id in the database.
-    Dictionary<int, int> m_LocalToDbFileId = new();
-
-    private Dictionary<string, ISQLiteHandler> m_Handlers = new()
-    {
-        { "Mesh", new MeshHandler() },
-        { "Texture2D", new Texture2DHandler() },
-        { "Shader", new ShaderHandler() },
-        { "AudioClip", new AudioClipHandler() },
-        { "AnimationClip", new AnimationClipHandler() },
-        { "AssetBundle", new AssetBundleHandler() },
-        { "PreloadData", new PreloadDataHandler() },
-    };
-
-    private SqliteConnection m_Database;
-    // serialized files
-    private AddReference m_AddReferenceCommand = new AddReference();
-    private AddAssetBundle m_AddAssetBundleCommand = new AddAssetBundle();
-    private AddSerializedFile m_AddSerializedFileCommand = new AddSerializedFile();
-    private AddObject m_AddObjectCommand = new AddObject();
-    private AddType m_AddTypeCommand = new AddType();
-    private AddAssetDependency m_InsertDepCommand = new AddAssetDependency();
-
-    private AddressablesBuild m_AddressablesBuild = new AddressablesBuild();
-    private AddressablesBuildBundle m_AddressablesBuildBundle = new AddressablesBuildBundle();
-    private AddressablesBuildBundleDependency m_AddressablesBuildBundleDependency = new AddressablesBuildBundleDependency();
-    private AddressablesBuildBundleExpandedDependency m_AddressablesBuildBundleExpandedDependency = new AddressablesBuildBundleExpandedDependency();
-    private AddressablesBuildBundleRegularDependency m_AddressablesBuildBundleRegularDependency = new AddressablesBuildBundleRegularDependency();
-    private AddressablesBuildBundleDependentBundle m_AddressablesBuildBundleDependentBundle = new AddressablesBuildBundleDependentBundle();
-    private AddressablesBuildBundleFile m_AddressablesBuildBundleFile = new AddressablesBuildBundleFile();
-
-    private AddressablesBuildDataFromOtherAsset m_AddressablesDataFromOtherAsset = new AddressablesBuildDataFromOtherAsset();
-    private AddressablesBuildDataFromOtherAssetObject m_AddressablesBuildDataFromOtherAssetObject = new AddressablesBuildDataFromOtherAssetObject();
-    private AddressablesBuildDataFromOtherAssetObjectReference m_AddressablesBuildDataFromOtherAssetObjectReference = new AddressablesBuildDataFromOtherAssetObjectReference();
-    private AddressablesBuildDataFromOtherAssetReferencingAsset m_AddressablesBuildDataFromOtherAssetReferencingAsset = new AddressablesBuildDataFromOtherAssetReferencingAsset();
-
-    private AddressablesBuildExplicitAsset m_AddressablesExplicitAsset = new AddressablesBuildExplicitAsset();
-    private AddressablesBuildExplicitAssetExternallyReferencedAsset m_AddressablesBuildExplicitAssetExternallyReferencedAsset = new AddressablesBuildExplicitAssetExternallyReferencedAsset();
-    private AddressablesBuildExplicitAssetInternalReferencedExplicitAsset m_AddressablesBuildExplicitAssetInternalReferencedExplicitAsset = new AddressablesBuildExplicitAssetInternalReferencedExplicitAsset();
-    private AddressablesBuildExplicitAssetInternalReferencedOtherAsset m_AddressablesBuildExplicitAssetInternalReferencedOtherAsset = new AddressablesBuildExplicitAssetInternalReferencedOtherAsset();
-    private AddressablesBuildExplicitAssetLabel m_AddressablesBuildExplicitAssetLabel = new AddressablesBuildExplicitAssetLabel();
-
-    private AddressablesBuildFile m_AddressablesBuildFile = new AddressablesBuildFile();
-    private AddressablesBuildFileAsset m_AddressablesBuildFileAsset = new AddressablesBuildFileAsset();
-    private AddressablesBuildFileExternalReference m_AddressablesBuildFileExternalReference = new AddressablesBuildFileExternalReference();
-    private AddressablesBuildFileOtherAsset m_AddressablesBuildFileOtherAsset = new AddressablesBuildFileOtherAsset();
-    private AddressablesBuildFileSubFile m_AddressablesBuildFileSubFile = new AddressablesBuildFileSubFile();
-
-    private AddressablesBuildGroup m_AddressablesBuildGroup = new AddressablesBuildGroup();
-    private AddressablesBuildGroupBundle m_AddressablesBuildGroupBundle = new AddressablesBuildGroupBundle();
-    private AddressablesBuildGroupSchema m_AddressablesBuildGroupSchema = new AddressablesBuildGroupSchema();
-
-    private AddressablesBuildSchema m_AddressablesBuildSchema = new AddressablesBuildSchema();
-    private AddressablesBuildSchemaDataPair m_AddressablesBuildSchemaDataPair = new AddressablesBuildSchemaDataPair();
-
-    private AddressablesBuildSubFile m_AddressablesBuildSubFile = new AddressablesBuildSubFile();
-    private SqliteCommand m_LastId = new SqliteCommand();
-    private SqliteTransaction m_CurrentTransaction = null;
-    public SQLiteWriter(string databaseName, bool skipReferences)
-    {
-        m_DatabaseName = databaseName;
-        m_SkipReferences = skipReferences;
-    }
-
-    public void Begin()
-    {
-        if (m_Database != null)
-        {
-            throw new InvalidOperationException("SQLiteWriter.Begin called twice");
-        }
-        SqliteConnectionStringBuilder builder = new();
-        builder.DataSource = m_DatabaseName;
-        builder.Mode = SqliteOpenMode.ReadWriteCreate;
-        m_Database = new SqliteConnection(builder.ConnectionString);
-        File.WriteAllBytes(m_DatabaseName, Array.Empty<byte>());
-        try
-        {
-            m_Database.Open();
-        }
-        catch (Exception e)
-        {
-            Console.Error.WriteLine($"Error creating database: {e.Message}");
-        }
-
-        using var command = m_Database.CreateCommand();
-        command.CommandText = Resources.Init;
-        command.ExecuteNonQuery();
-
-        foreach (var handler in m_Handlers.Values)
-        {
-            // Console.WriteLine($"Init handler: {handler.GetType().Name}");
-            // Console.WriteLine($"Connection state before init: {m_Database.State}");
-            handler.Init(m_Database);
-            // Console.WriteLine($"Connection state after init: {m_Database.State}");
-
-        }
-
-        CreateSQLiteCommands();
-    }
-
-    public void End()
-    {
-        if (m_Database == null)
-        {
-            throw new InvalidOperationException("SQLiteWriter.End called before SQLiteWriter.Begin");
-        }
-
-        foreach (var handler in m_Handlers.Values)
-        {
-            handler.Finalize(m_Database);
-        }
-
-        using var finalizeCommand = m_Database.CreateCommand();
-        finalizeCommand.CommandText = Resources.Finalize;
-        finalizeCommand.ExecuteNonQuery();
-    }
-
-    private void CreateSQLiteCommands()
-    {
-
-        // build serialized file commands
-        m_AddReferenceCommand.CreateCommand(m_Database);
-        m_AddAssetBundleCommand.CreateCommand(m_Database);
-        m_AddSerializedFileCommand.CreateCommand(m_Database);
-        m_AddObjectCommand.CreateCommand(m_Database);
-        m_AddTypeCommand.CreateCommand(m_Database);
-        m_InsertDepCommand.CreateCommand(m_Database);
-
-        // build addressables file commands
-        m_AddressablesBuild.CreateCommand(m_Database);
-        // Build Bundle Tables
-        m_AddressablesBuildBundle.CreateCommand(m_Database);
-        m_AddressablesBuildBundleDependency.CreateCommand(m_Database);
-        m_AddressablesBuildBundleExpandedDependency.CreateCommand(m_Database);
-        m_AddressablesBuildBundleRegularDependency.CreateCommand(m_Database);
-        m_AddressablesBuildBundleDependentBundle.CreateCommand(m_Database);
-        m_AddressablesBuildBundleFile.CreateCommand(m_Database);
-
-        // Data From Other Asset Tables
-        // Data From Other Asset Tables
-        m_AddressablesDataFromOtherAsset.CreateCommand(m_Database);
-        m_AddressablesBuildDataFromOtherAssetObject.CreateCommand(m_Database);
-        m_AddressablesBuildDataFromOtherAssetObjectReference.CreateCommand(m_Database);
-        m_AddressablesBuildDataFromOtherAssetReferencingAsset.CreateCommand(m_Database);
-
-        // Explicit Asset Tables
-        m_AddressablesExplicitAsset.CreateCommand(m_Database);
-        m_AddressablesBuildExplicitAssetExternallyReferencedAsset.CreateCommand(m_Database);
-        m_AddressablesBuildExplicitAssetInternalReferencedExplicitAsset.CreateCommand(m_Database);
-        m_AddressablesBuildExplicitAssetInternalReferencedOtherAsset.CreateCommand(m_Database);
-        m_AddressablesBuildExplicitAssetLabel.CreateCommand(m_Database);
-
-        // File Tables
-        m_AddressablesBuildFile.CreateCommand(m_Database);
-        m_AddressablesBuildFileAsset.CreateCommand(m_Database);
-        m_AddressablesBuildFileExternalReference.CreateCommand(m_Database);
-        m_AddressablesBuildFileOtherAsset.CreateCommand(m_Database);
-        m_AddressablesBuildFileSubFile.CreateCommand(m_Database);
-
-        // Group Tables
-        m_AddressablesBuildGroup.CreateCommand(m_Database);
-        m_AddressablesBuildGroupBundle.CreateCommand(m_Database);
-        m_AddressablesBuildGroupSchema.CreateCommand(m_Database);
-
-        // Schema Tables
-        m_AddressablesBuildSchema.CreateCommand(m_Database);
-        m_AddressablesBuildSchemaDataPair.CreateCommand(m_Database);
-
-        // SubFile Tables
-        m_AddressablesBuildSubFile.CreateCommand(m_Database);
-
-        m_LastId = m_Database.CreateCommand();
-        m_LastId.CommandText = "SELECT last_insert_rowid()";
-    }
-    public void BeginAssetBundle(string name, long size)
-    {
-        if (m_CurrentAssetBundleId != -1)
-        {
-            throw new InvalidOperationException("SQLWriter.BeginAssetBundle called twice");
-        }
-
-        m_CurrentAssetBundleId = m_NextAssetBundleId++;
-        m_AddAssetBundleCommand.SetValue("id", m_CurrentAssetBundleId);
-        m_AddAssetBundleCommand.SetValue("name", name);
-        m_AddAssetBundleCommand.SetValue("file_size", size);
-        m_AddAssetBundleCommand.ExecuteNonQuery();
-    }
-
-    public void EndAssetBundle()
-    {
-        if (m_CurrentAssetBundleId == -1)
-        {
-            throw new InvalidOperationException("SQLWriter.EndAssetBundle called before SQLWriter.BeginAssetBundle");
-        }
-
-        m_CurrentAssetBundleId = -1;
-    }
-
-    public void WriteAddressablesBuild(string filename, BuildLayout buildLayout)
-    {
-        using var transaction = m_Database.BeginTransaction();
-        m_CurrentTransaction = transaction;
-
-        try
-        {
-            m_AddressablesBuild.SetTransaction(transaction);
-            m_AddressablesBuild.SetValue("name", Path.GetFileName(filename));
-            m_AddressablesBuild.SetValue("build_target", buildLayout.BuildTarget);
-            m_AddressablesBuild.SetValue("start_time", buildLayout.BuildStartTime);
-            m_AddressablesBuild.SetValue("duration", buildLayout.Duration);
-            m_AddressablesBuild.SetValue("error", buildLayout.BuildError);
-            m_AddressablesBuild.SetValue("package_version", buildLayout.PackageVersion);
-            m_AddressablesBuild.SetValue("player_version", buildLayout.PlayerBuildVersion);
-            m_AddressablesBuild.SetValue("build_script", buildLayout.BuildScript);
-            m_AddressablesBuild.SetValue("result_hash", buildLayout.BuildResultHash);
-            m_AddressablesBuild.SetValue("type", buildLayout.BuildType);
-            m_AddressablesBuild.SetValue("unity_version", buildLayout.UnityVersion);
-            m_AddressablesBuild.ExecuteNonQuery();
-
-            m_LastId.Transaction = transaction;
-            long buildId = (long)m_LastId.ExecuteScalar();
-            Console.WriteLine($"Build ID: {buildId}");
-
-            foreach (var reference in buildLayout.references.RefIds)
-            {
-                switch (reference.type.Class)
-                {
-                    case "BuildLayout/Bundle":
-                        WriteBuildLayoutBundle(reference, buildId, transaction);
-                        break;
-
-                    case "BuildLayout/DataFromOtherAsset":
-                        WriteBuildLayoutDataFromOtherAsset(reference, buildId, transaction);
-                        break;
-
-                    case "BuildLayout/ExplicitAsset":
-                        WriteBuildLayoutExplicitAsset(reference, buildId, transaction);
-                        break;
-
-                    case "BuildLayout/File":
-                        WriteBuildLayoutFile(reference, buildId, transaction);
-                        break;
-
-                    case "BuildLayout/Group":
-                        WriteBuildLayoutGroup(reference, buildId, transaction);
-                        break;
-
-                    case "BuildLayout/SchemaData":
-                        WriteBuildLayoutSchemaData(reference, buildId, transaction);
-                        break;
-
-                    case "BuildLayout/SubFile":
-                        WriteBuildLayoutSubFile(reference, buildId, transaction);
-                        break;
-                }
-            }
-
-            // do the stuff
-            transaction.Commit();
-        }
-        catch (Exception e)
-        {
-            Console.WriteLine(e.StackTrace.ToString());
-            transaction.Rollback();
-            throw;
-        }
-    }
-
-    private void WriteBuildLayoutBundle(Reference reference, long buildId, SqliteTransaction transaction)
-    {
-        m_AddressablesBuildBundle.SetTransaction(transaction);
-        m_AddressablesBuildBundle.SetValue("id", reference.rid);
-        m_AddressablesBuildBundle.SetValue("build_id", buildId);
-        m_AddressablesBuildBundle.SetValue("asset_count", reference.data.AssetCount);
-        m_AddressablesBuildBundle.SetValue("build_status", reference.data.BuildStatus);
-        m_AddressablesBuildBundle.SetValue("crc", reference.data.CRC);
-        m_AddressablesBuildBundle.SetValue("compression", reference.data.Compression);
-        m_AddressablesBuildBundle.SetValue("dependency_file_size", reference.data.DependencyFileSize);
-        m_AddressablesBuildBundle.SetValue("expanded_dependency_file_size", reference.data.ExpandedDependencyFileSize);
-        m_AddressablesBuildBundle.SetValue("file_size", reference.data.FileSize);
-        m_AddressablesBuildBundle.SetValue("group_rid", reference.data.Group.rid);
-        m_AddressablesBuildBundle.SetValue("hash", JsonConvert.SerializeObject(reference.data.Hash));
-        m_AddressablesBuildBundle.SetValue("internal_name", reference.data.InternalName);
-        m_AddressablesBuildBundle.SetValue("load_path", reference.data.LoadPath);
-        m_AddressablesBuildBundle.SetValue("name", reference.data.Name);
-        m_AddressablesBuildBundle.SetValue("provider", reference.data.Provider);
-        m_AddressablesBuildBundle.SetValue("result_type", reference.data.ResultType);
-        m_AddressablesBuildBundle.ExecuteNonQuery();
-
-        // Insert bundle dependencies
-        if (reference.data.BundleDependencies != null)
-        {
-            foreach (var dep in reference.data.BundleDependencies)
-            {
-                m_AddressablesBuildBundleDependency.SetTransaction(transaction);
-                m_AddressablesBuildBundleDependency.SetValue("bundle_id", reference.rid);
-                m_AddressablesBuildBundleDependency.SetValue("build_id", buildId);
-                m_AddressablesBuildBundleDependency.SetValue("dependency_rid", dep.rid);
-                m_AddressablesBuildBundleDependency.ExecuteNonQuery();
-            }
-        }
-
-        // Insert regular dependencies
-        if (reference.data.Dependencies != null)
-        {
-            foreach (var dep in reference.data.Dependencies)
-            {
-                m_AddressablesBuildBundleRegularDependency.SetTransaction(transaction);
-                m_AddressablesBuildBundleRegularDependency.SetValue("bundle_id", reference.rid);
-                m_AddressablesBuildBundleRegularDependency.SetValue("build_id", buildId);
-                m_AddressablesBuildBundleRegularDependency.SetValue("dependency_rid", dep.rid);
-                m_AddressablesBuildBundleRegularDependency.ExecuteNonQuery();
-            }
-        }
-
-        // Insert dependent bundles
-        if (reference.data.DependentBundles != null)
-        {
-            foreach (var depBundle in reference.data.DependentBundles)
-            {
-                m_AddressablesBuildBundleDependentBundle.SetTransaction(transaction);
-                m_AddressablesBuildBundleDependentBundle.SetValue("bundle_id", reference.rid);
-                m_AddressablesBuildBundleDependentBundle.SetValue("build_id", buildId);
-                m_AddressablesBuildBundleDependentBundle.SetValue("dependent_bundle_rid", depBundle.rid);
-                m_AddressablesBuildBundleDependentBundle.ExecuteNonQuery();
-            }
-        }
-
-        // Insert expanded dependencies
-        if (reference.data.ExpandedDependencies != null)
-        {
-            foreach (var dep in reference.data.ExpandedDependencies)
-            {
-                m_AddressablesBuildBundleExpandedDependency.SetTransaction(transaction);
-                m_AddressablesBuildBundleExpandedDependency.SetValue("bundle_id", reference.rid);
-                m_AddressablesBuildBundleExpandedDependency.SetValue("build_id", buildId);
-                m_AddressablesBuildBundleExpandedDependency.SetValue("dependency_rid", dep.rid);
-                m_AddressablesBuildBundleExpandedDependency.ExecuteNonQuery();
-            }
-        }
-
-        // Insert files
-        if (reference.data.Files != null)
-        {
-            foreach (var file in reference.data.Files)
-            {
-                m_AddressablesBuildBundleFile.SetTransaction(transaction);
-                m_AddressablesBuildBundleFile.SetValue("bundle_id", reference.rid);
-                m_AddressablesBuildBundleFile.SetValue("build_id", buildId);
-                m_AddressablesBuildBundleFile.SetValue("file_rid", file.rid);
-                m_AddressablesBuildBundleFile.ExecuteNonQuery();
-            }
-        }
-    }
-
-    private void WriteBuildLayoutDataFromOtherAsset(Reference reference, long buildId, SqliteTransaction transaction)
-    {
-        m_AddressablesDataFromOtherAsset.SetTransaction(transaction);
-        m_AddressablesDataFromOtherAsset.SetValue("id", reference.rid);
-        m_AddressablesDataFromOtherAsset.SetValue("build_id", buildId);
-        m_AddressablesDataFromOtherAsset.SetValue("asset_guid", reference.data.AssetGuid);
-        m_AddressablesDataFromOtherAsset.SetValue("asset_path", reference.data.AssetPath);
-        m_AddressablesDataFromOtherAsset.SetValue("file", reference.data.File.rid);
-        m_AddressablesDataFromOtherAsset.SetValue("main_asset_type", reference.data.MainAssetType);
-        m_AddressablesDataFromOtherAsset.SetValue("object_count", reference.data.ObjectCount);
-        m_AddressablesDataFromOtherAsset.SetValue("serialized_size", reference.data.SerializedSize);
-        m_AddressablesDataFromOtherAsset.SetValue("streamed_size", reference.data.StreamedSize);
-        m_AddressablesDataFromOtherAsset.ExecuteNonQuery();
-
-        // Insert objects
-        if (reference.data.Objects != null)
-        {
-            foreach (var obj in reference.data.Objects)
-            {
-                m_AddressablesBuildDataFromOtherAssetObject.SetTransaction(transaction);
-                m_AddressablesBuildDataFromOtherAssetObject.SetValue("data_from_other_asset_id", reference.rid);
-                m_AddressablesBuildDataFromOtherAssetObject.SetValue("build_id", buildId);
-                m_AddressablesBuildDataFromOtherAssetObject.SetValue("asset_type", obj.AssetType);
-                m_AddressablesBuildDataFromOtherAssetObject.SetValue("component_name", obj.ComponentName ?? "");
-                m_AddressablesBuildDataFromOtherAssetObject.SetValue("local_identifier_in_file", obj.LocalIdentifierInFile);
-                m_AddressablesBuildDataFromOtherAssetObject.SetValue("object_name", obj.ObjectName ?? "");
-                m_AddressablesBuildDataFromOtherAssetObject.SetValue("serialized_size", obj.SerializedSize);
-                m_AddressablesBuildDataFromOtherAssetObject.SetValue("streamed_size", obj.StreamedSize);
-                m_AddressablesBuildDataFromOtherAssetObject.ExecuteNonQuery();
-
-                // Insert object references
-                if (obj.References != null)
-                {
-                    foreach (var objRef in obj.References)
-                    {
-                        m_AddressablesBuildDataFromOtherAssetObjectReference.SetTransaction(transaction);
-                        m_AddressablesBuildDataFromOtherAssetObjectReference.SetValue("data_from_other_asset_id", reference.rid);
-                        m_AddressablesBuildDataFromOtherAssetObjectReference.SetValue("build_id", buildId);
-                        m_AddressablesBuildDataFromOtherAssetObjectReference.SetValue("local_identifier_in_file", obj.LocalIdentifierInFile);
-                        m_AddressablesBuildDataFromOtherAssetObjectReference.SetValue("asset_id", objRef.AssetId);
-                        m_AddressablesBuildDataFromOtherAssetObjectReference.SetValue("object_id", objRef.ObjectId);
-                        m_AddressablesBuildDataFromOtherAssetObjectReference.ExecuteNonQuery();
-                    }
-                }
-            }
-        }
-
-        // Insert referencing assets
-        if (reference.data.ReferencingAssets != null)
-        {
-            foreach (var refAsset in reference.data.ReferencingAssets)
-            {
-                m_AddressablesBuildDataFromOtherAssetReferencingAsset.SetTransaction(transaction);
-                m_AddressablesBuildDataFromOtherAssetReferencingAsset.SetValue("data_from_other_asset_id", reference.rid);
-                m_AddressablesBuildDataFromOtherAssetReferencingAsset.SetValue("build_id", buildId);
-                m_AddressablesBuildDataFromOtherAssetReferencingAsset.SetValue("referencing_asset_rid", refAsset.rid);
-                m_AddressablesBuildDataFromOtherAssetReferencingAsset.ExecuteNonQuery();
-            }
-        }
-    }
-
-    private void WriteBuildLayoutExplicitAsset(Reference reference, long buildId, SqliteTransaction transaction)
-    {
-        m_AddressablesExplicitAsset.SetTransaction(transaction);
-        m_AddressablesExplicitAsset.SetValue("id", reference.rid);
-        m_AddressablesExplicitAsset.SetValue("build_id", buildId);
-        m_AddressablesExplicitAsset.SetValue("bundle", reference.data.Bundle.rid);
-        m_AddressablesExplicitAsset.SetValue("file", reference.data.File.rid);
-        m_AddressablesExplicitAsset.SetValue("asset_hash", reference.data.AssetHash.Hash);
-        m_AddressablesExplicitAsset.SetValue("asset_path", reference.data.AssetPath);
-        m_AddressablesExplicitAsset.SetValue("addressable_name", reference.data.AddressableName);
-        m_AddressablesExplicitAsset.SetValue("group_guid", reference.data.GroupGuid);
-        m_AddressablesExplicitAsset.SetValue("guid", reference.data.Guid);
-        m_AddressablesExplicitAsset.SetValue("internal_id", reference.data.InternalId);
-        m_AddressablesExplicitAsset.SetValue("streamed_size", reference.data.StreamedSize);
-        m_AddressablesExplicitAsset.SetValue("serialized_size", reference.data.SerializedSize);
-        m_AddressablesExplicitAsset.SetValue("main_asset_type", reference.data.MainAssetType);
-        m_AddressablesExplicitAsset.ExecuteNonQuery();
-
-        // Insert externally referenced assets
-        if (reference.data.ExternallyReferencedAssets != null)
-        {
-            foreach (var extRefAsset in reference.data.ExternallyReferencedAssets)
-            {
-                m_AddressablesBuildExplicitAssetExternallyReferencedAsset.SetTransaction(transaction);
-                m_AddressablesBuildExplicitAssetExternallyReferencedAsset.SetValue("explicit_asset_id", reference.rid);
-                m_AddressablesBuildExplicitAssetExternallyReferencedAsset.SetValue("build_id", buildId);
-                m_AddressablesBuildExplicitAssetExternallyReferencedAsset.SetValue("externally_referenced_asset_rid", extRefAsset.rid);
-                m_AddressablesBuildExplicitAssetExternallyReferencedAsset.ExecuteNonQuery();
-            }
-        }
-
-        // Insert internal referenced explicit assets
-        if (reference.data.InternalReferencedExplicitAssets != null)
-        {
-            foreach (var intRefExplicitAsset in reference.data.InternalReferencedExplicitAssets)
-            {
-                m_AddressablesBuildExplicitAssetInternalReferencedExplicitAsset.SetTransaction(transaction);
-                m_AddressablesBuildExplicitAssetInternalReferencedExplicitAsset.SetValue("explicit_asset_id", reference.rid);
-                m_AddressablesBuildExplicitAssetInternalReferencedExplicitAsset.SetValue("build_id", buildId);
-                m_AddressablesBuildExplicitAssetInternalReferencedExplicitAsset.SetValue("internal_referenced_explicit_asset_rid", intRefExplicitAsset.rid);
-                m_AddressablesBuildExplicitAssetInternalReferencedExplicitAsset.ExecuteNonQuery();
-            }
-        }
-
-        // Insert internal referenced other assets
-        if (reference.data.InternalReferencedOtherAssets != null)
-        {
-            foreach (var intRefOtherAsset in reference.data.InternalReferencedOtherAssets)
-            {
-                m_AddressablesBuildExplicitAssetInternalReferencedOtherAsset.SetTransaction(transaction);
-                m_AddressablesBuildExplicitAssetInternalReferencedOtherAsset.SetValue("explicit_asset_id", reference.rid);
-                m_AddressablesBuildExplicitAssetInternalReferencedOtherAsset.SetValue("build_id", buildId);
-                m_AddressablesBuildExplicitAssetInternalReferencedOtherAsset.SetValue("internal_referenced_other_asset_rid", intRefOtherAsset.rid);
-                m_AddressablesBuildExplicitAssetInternalReferencedOtherAsset.ExecuteNonQuery();
-            }
-        }
-
-        // Insert labels
-        if (reference.data.Labels != null)
-        {
-            foreach (var label in reference.data.Labels)
-            {
-                m_AddressablesBuildExplicitAssetLabel.SetTransaction(transaction);
-                m_AddressablesBuildExplicitAssetLabel.SetValue("explicit_asset_id", reference.rid);
-                m_AddressablesBuildExplicitAssetLabel.SetValue("build_id", buildId);
-                m_AddressablesBuildExplicitAssetLabel.SetValue("label", label);
-                m_AddressablesBuildExplicitAssetLabel.ExecuteNonQuery();
-            }
-        }
-    }
-
-    private void WriteBuildLayoutFile(Reference reference, long buildId, SqliteTransaction transaction)
-    {
-        m_AddressablesBuildFile.SetTransaction(transaction);
-        m_AddressablesBuildFile.SetValue("id", reference.rid);
-        m_AddressablesBuildFile.SetValue("build_id", buildId);
-        m_AddressablesBuildFile.SetValue("bundle", reference.data.Bundle.rid);
-        m_AddressablesBuildFile.SetValue("bundle_object_info_size", reference.data.BundleObjectInfo.Size);
-        m_AddressablesBuildFile.SetValue("mono_script_count", reference.data.MonoScriptCount);
-        m_AddressablesBuildFile.SetValue("mono_script_size", reference.data.MonoScriptSize);
-        m_AddressablesBuildFile.SetValue("name", reference.data.Name);
-        m_AddressablesBuildFile.SetValue("preload_info_size", reference.data.PreloadInfoSize);
-        m_AddressablesBuildFile.SetValue("write_result_filename", reference.data.WriteResultFilename);
-        m_AddressablesBuildFile.ExecuteNonQuery();
-
-        // Insert assets
-        if (reference.data.Assets != null)
-        {
-            foreach (var asset in reference.data.Assets)
-            {
-                m_AddressablesBuildFileAsset.SetTransaction(transaction);
-                m_AddressablesBuildFileAsset.SetValue("file_id", reference.rid);
-                m_AddressablesBuildFileAsset.SetValue("build_id", buildId);
-                m_AddressablesBuildFileAsset.SetValue("asset_rid", asset.rid);
-                m_AddressablesBuildFileAsset.ExecuteNonQuery();
-            }
-        }
-
-        // Insert external references
-        if (reference.data.ExternalReferences != null)
-        {
-            foreach (var extRef in reference.data.ExternalReferences)
-            {
-                m_AddressablesBuildFileExternalReference.SetTransaction(transaction);
-                m_AddressablesBuildFileExternalReference.SetValue("file_id", reference.rid);
-                m_AddressablesBuildFileExternalReference.SetValue("build_id", buildId);
-                m_AddressablesBuildFileExternalReference.SetValue("external_reference_rid", extRef.rid);
-                m_AddressablesBuildFileExternalReference.ExecuteNonQuery();
-            }
-        }
-
-        // Insert other assets
-        if (reference.data.OtherAssets != null)
-        {
-            foreach (var otherAsset in reference.data.OtherAssets)
-            {
-                m_AddressablesBuildFileOtherAsset.SetTransaction(transaction);
-                m_AddressablesBuildFileOtherAsset.SetValue("file_id", reference.rid);
-                m_AddressablesBuildFileOtherAsset.SetValue("build_id", buildId);
-                m_AddressablesBuildFileOtherAsset.SetValue("other_asset_rid", otherAsset.rid);
-                m_AddressablesBuildFileOtherAsset.ExecuteNonQuery();
-            }
-        }
-
-        // Insert sub files
-        if (reference.data.SubFiles != null)
-        {
-            foreach (var subFile in reference.data.SubFiles)
-            {
-                m_AddressablesBuildFileSubFile.SetTransaction(transaction);
-                m_AddressablesBuildFileSubFile.SetValue("file_id", reference.rid);
-                m_AddressablesBuildFileSubFile.SetValue("build_id", buildId);
-                m_AddressablesBuildFileSubFile.SetValue("sub_file_rid", subFile.rid);
-                m_AddressablesBuildFileSubFile.ExecuteNonQuery();
-            }
-        }
-    }
-
-    private void WriteBuildLayoutGroup(Reference reference, long buildId, SqliteTransaction transaction)
-    {
-        m_AddressablesBuildGroup.SetTransaction(transaction);
-        m_AddressablesBuildGroup.SetValue("id", reference.rid);
-        m_AddressablesBuildGroup.SetValue("build_id", buildId);
-        m_AddressablesBuildGroup.SetValue("guid", reference.data.Guid);
-        m_AddressablesBuildGroup.SetValue("name", reference.data.Name);
-        m_AddressablesBuildGroup.SetValue("packing_mode", reference.data.PackingMode);
-        m_AddressablesBuildGroup.ExecuteNonQuery();
-
-        // Insert bundles
-        if (reference.data.Bundles != null)
-        {
-            foreach (var bundle in reference.data.Bundles)
-            {
-                m_AddressablesBuildGroupBundle.SetTransaction(transaction);
-                m_AddressablesBuildGroupBundle.SetValue("group_id", reference.rid);
-                m_AddressablesBuildGroupBundle.SetValue("build_id", buildId);
-                m_AddressablesBuildGroupBundle.SetValue("bundle_rid", bundle.rid);
-                m_AddressablesBuildGroupBundle.ExecuteNonQuery();
-            }
-        }
-
-        // Insert schemas
-        if (reference.data.Schemas != null)
-        {
-            foreach (var schema in reference.data.Schemas)
-            {
-                m_AddressablesBuildGroupSchema.SetTransaction(transaction);
-                m_AddressablesBuildGroupSchema.SetValue("group_id", reference.rid);
-                m_AddressablesBuildGroupSchema.SetValue("build_id", buildId);
-                m_AddressablesBuildGroupSchema.SetValue("schema_rid", schema.rid);
-                m_AddressablesBuildGroupSchema.ExecuteNonQuery();
-            }
-        }
-    }
-
-    private void WriteBuildLayoutSchemaData(Reference reference, long buildId, SqliteTransaction transaction)
-    {
-        m_AddressablesBuildSchema.SetTransaction(transaction);
-        m_AddressablesBuildSchema.SetValue("id", reference.rid);
-        m_AddressablesBuildSchema.SetValue("build_id", buildId);
-        m_AddressablesBuildSchema.SetValue("guid", reference.data.Guid);
-        m_AddressablesBuildSchema.SetValue("type", reference.data.Type);
-        m_AddressablesBuildSchema.ExecuteNonQuery();
-
-        // Insert schema data pairs
-        if (reference.data.SchemaDataPairs != null)
-        {
-            foreach (var dataPair in reference.data.SchemaDataPairs)
-            {
-                m_AddressablesBuildSchemaDataPair.SetTransaction(transaction);
-                m_AddressablesBuildSchemaDataPair.SetValue("schema_id", reference.rid);
-                m_AddressablesBuildSchemaDataPair.SetValue("build_id", buildId);
-                m_AddressablesBuildSchemaDataPair.SetValue("key", dataPair.Key);
-                m_AddressablesBuildSchemaDataPair.SetValue("value", dataPair.Value);
-                m_AddressablesBuildSchemaDataPair.ExecuteNonQuery();
-            }
-        }
-    }
-
-    private void WriteBuildLayoutSubFile(Reference reference, long buildId, SqliteTransaction transaction)
-    {
-        m_AddressablesBuildSubFile.SetTransaction(transaction);
-        m_AddressablesBuildSubFile.SetValue("id", reference.rid);
-        m_AddressablesBuildSubFile.SetValue("build_id", buildId);
-        m_AddressablesBuildSubFile.SetValue("is_serialized_file", reference.data.IsSerializedFile ? 1 : 0);
-        m_AddressablesBuildSubFile.SetValue("name", reference.data.Name);
-        m_AddressablesBuildSubFile.SetValue("size", reference.data.Size);
-        m_AddressablesBuildSubFile.ExecuteNonQuery();
-    }
-
-    public void WriteSerializedFile(string relativePath, string fullPath, string containingFolder)
-    {
-        using var sf = UnityFileSystem.OpenSerializedFile(fullPath);
-        using var reader = new UnityFileReader(fullPath, 64 * 1024 * 1024);
-        using var pptrReader = new PPtrAndCrcProcessor(sf, reader, containingFolder, AddReference);
-        int serializedFileId = m_SerializedFileIdProvider.GetId(Path.GetFileName(fullPath).ToLower());
-        int sceneId = -1;
-
-        using var transaction = m_Database.BeginTransaction();
-        m_CurrentTransaction = transaction;
-
-        var match = m_RegexSceneFile.Match(relativePath);
-
-        if (match.Success)
-        {
-            var sceneName = match.Groups[1].Value;
-
-            // There is no Scene object in Unity (a Scene is the full content of a
-            // SerializedFile). We generate an object id using the name of the Scene
-            // as SerializedFile name, and the object id 0.
-            sceneId = m_ObjectIdProvider.GetId((m_SerializedFileIdProvider.GetId(sceneName), 0));
-
-            // There are 2 SerializedFiles per Scene, one ends with .sharedAssets. This is a
-            // dirty trick to avoid inserting the scene object a second time.
-            if (relativePath.EndsWith(".sharedAssets"))
-            {
-                m_AddObjectCommand.SetTransaction(transaction);
-                m_AddObjectCommand.SetValue("game_object", ""); // or other value
-                m_AddObjectCommand.SetValue("id", sceneId);
-                m_AddObjectCommand.SetValue("object_id", 0);
-                m_AddObjectCommand.SetValue("serialized_file", serializedFileId);
-                // The type is set to -1 which doesn't exist in Unity, but is associated to
-                // "Scene" in the database.
-                m_AddObjectCommand.SetValue("type", -1);
-                m_AddObjectCommand.SetValue("name", sceneName);
-                m_AddObjectCommand.SetValue("size", 0);
-                m_AddObjectCommand.SetValue("crc32", 0);
-                m_AddObjectCommand.ExecuteNonQuery();
-            }
-        }
-
-        m_LocalToDbFileId.Clear();
-
-        Context ctx = new()
-        {
-            AssetBundleId = m_CurrentAssetBundleId,
-            SerializedFileId = serializedFileId,
-            SceneId = sceneId,
-            ObjectIdProvider = m_ObjectIdProvider,
-            SerializedFileIdProvider = m_SerializedFileIdProvider,
-            LocalToDbFileId = m_LocalToDbFileId,
-            Transaction = transaction,
-        };
-
-        ctx.Transaction = transaction;
-        try
-        {
-            m_AddSerializedFileCommand.SetTransaction(transaction);
-            m_AddSerializedFileCommand.SetValue("id", serializedFileId);
-            m_AddSerializedFileCommand.SetValue("asset_bundle", m_CurrentAssetBundleId == -1 ? "" : m_CurrentAssetBundleId);
-            m_AddSerializedFileCommand.SetValue("name", relativePath);
-            m_AddSerializedFileCommand.ExecuteNonQuery();
-
-            int localId = 0;
-            m_LocalToDbFileId.Add(localId++, serializedFileId);
-            foreach (var extRef in sf.ExternalReferences)
-            {
-                m_LocalToDbFileId.Add(localId++,
-                    m_SerializedFileIdProvider.GetId(extRef.Path.Substring(extRef.Path.LastIndexOf('/') + 1).ToLower()));
-            }
-
-            foreach (var obj in sf.Objects)
-            {
-                var currentObjectId = m_ObjectIdProvider.GetId((serializedFileId, obj.Id));
-                // Console.WriteLine($"\nProcessing {currentObjectId}");
-                var root = sf.GetTypeTreeRoot(obj.Id);
-                var offset = obj.Offset;
-                uint crc32 = 0;
-
-                if (!m_TypeSet.Contains(obj.TypeId))
-                {
-                    m_AddTypeCommand.SetTransaction(transaction);
-                    m_AddTypeCommand.SetValue("id", obj.TypeId);
-                    m_AddTypeCommand.SetValue("name", root.Type);
-                    m_AddTypeCommand.ExecuteNonQuery();
-
-                    m_TypeSet.Add(obj.TypeId);
-                }
-
-                var randomAccessReader = new RandomAccessReader(sf, root, reader, offset);
-
-                string name = string.Empty;
-                long streamDataSize = 0;
-
-                if (m_Handlers.TryGetValue(root.Type, out var handler))
-                {
-                    handler.Process(ctx, currentObjectId, randomAccessReader,
-                        out name, out streamDataSize);
-                }
-                else if (randomAccessReader.HasChild("m_Name"))
-                {
-                    name = randomAccessReader["m_Name"].GetValue<string>();
-                }
-
-                if (randomAccessReader.HasChild("m_GameObject"))
-                {
-                    var pptr = randomAccessReader["m_GameObject"];
-                    var fileId = m_LocalToDbFileId[pptr["m_FileID"].GetValue<int>()];
-                    var gameObjectID = m_ObjectIdProvider.GetId((fileId, pptr["m_PathID"].GetValue<long>()));
-                    m_AddObjectCommand.SetValue("game_object", gameObjectID);
-                }
-                else
-                {
-                    m_AddObjectCommand.SetValue("game_object", "");
-                }
-
-                if (!m_SkipReferences)
-                {
-                    crc32 = pptrReader.Process(currentObjectId, offset, root);
-                }
-
-                // convert this to the new syntax
-                m_AddObjectCommand.SetTransaction(transaction);
-                m_AddObjectCommand.SetValue("id", currentObjectId);
-                m_AddObjectCommand.SetValue("object_id", obj.Id);
-                m_AddObjectCommand.SetValue("serialized_file", serializedFileId);
-                m_AddObjectCommand.SetValue("type", obj.TypeId);
-                m_AddObjectCommand.SetValue("name", name);
-                m_AddObjectCommand.SetValue("size", obj.Size + streamDataSize);
-                m_AddObjectCommand.SetValue("crc32", crc32);
-                m_AddObjectCommand.ExecuteNonQuery();
-
-                // If this is a Scene AssetBundle, add the object as a depencency of the
-                // current scene.
-                if (ctx.SceneId != -1)
-                {
-                    m_InsertDepCommand.SetTransaction(transaction);
-                    m_InsertDepCommand.SetValue("object", ctx.SceneId);
-                    m_InsertDepCommand.SetValue("dependency", currentObjectId);
-                    m_InsertDepCommand.ExecuteNonQuery();
-                }
-            }
-
-            transaction.Commit();
-        }
-        catch (Exception)
-        {
-            transaction.Rollback();
-            throw;
-        }
-    }
-
-    private int AddReference(long objectId, int fileId, long pathId, string propertyPath, string propertyType)
-    {
-        var referencedObjectId = m_ObjectIdProvider.GetId((m_LocalToDbFileId[fileId], pathId));
-        m_AddReferenceCommand.SetTransaction(m_CurrentTransaction);
-        m_AddReferenceCommand.SetValue("object", objectId);
-        m_AddReferenceCommand.SetValue("referenced_object", referencedObjectId);
-        m_AddReferenceCommand.SetValue("property_path", propertyPath);
-        m_AddReferenceCommand.SetValue("property_type", propertyType);
-        m_AddReferenceCommand.ExecuteNonQuery();
-        return referencedObjectId;
-    }
-
-    public void Dispose()
-    {
-        foreach (var handler in m_Handlers.Values)
-        {
-            handler.Dispose();
-        }
-
-        // Serialized file dispose calls
-        m_AddAssetBundleCommand.Dispose();
-        m_AddSerializedFileCommand.Dispose();
-        m_AddReferenceCommand.Dispose();
-        m_AddObjectCommand.Dispose();
-        m_AddTypeCommand.Dispose();
-        m_InsertDepCommand.Dispose();
-
-        // Addressables dispose calls
-        m_AddressablesBuild.Dispose();
-        m_AddressablesBuildBundle.Dispose();
-        m_AddressablesBuildBundleDependency.Dispose();
-        m_AddressablesBuildBundleExpandedDependency.Dispose();
-        m_AddressablesBuildBundleRegularDependency.Dispose();
-        m_AddressablesBuildBundleDependentBundle.Dispose();
-        m_AddressablesBuildBundleFile.Dispose();
-        m_AddressablesDataFromOtherAsset.Dispose();
-        m_AddressablesBuildDataFromOtherAssetObject.Dispose();
-        m_AddressablesBuildDataFromOtherAssetObjectReference.Dispose();
-        m_AddressablesBuildDataFromOtherAssetReferencingAsset.Dispose();
-        m_AddressablesExplicitAsset.Dispose();
-        m_AddressablesBuildExplicitAssetExternallyReferencedAsset.Dispose();
-        m_AddressablesBuildExplicitAssetInternalReferencedExplicitAsset.Dispose();
-        m_AddressablesBuildExplicitAssetInternalReferencedOtherAsset.Dispose();
-        m_AddressablesBuildExplicitAssetLabel.Dispose();
-        m_AddressablesBuildFile.Dispose();
-        m_AddressablesBuildFileAsset.Dispose();
-        m_AddressablesBuildFileExternalReference.Dispose();
-        m_AddressablesBuildFileOtherAsset.Dispose();
-        m_AddressablesBuildFileSubFile.Dispose();
-        m_AddressablesBuildGroup.Dispose();
-        m_AddressablesBuildGroupBundle.Dispose();
-        m_AddressablesBuildGroupSchema.Dispose();
-        m_AddressablesBuildSchema.Dispose();
-        m_AddressablesBuildSchemaDataPair.Dispose();
-        m_AddressablesBuildSubFile.Dispose();
-        m_LastId.Dispose();
-        m_Database.Dispose();
-    }
-}
